improve the build script

+++ build.sh
@@ -1,7 +1,11 @@
 #!/usr/bin/env bash
 
+if [ -z "$1" ]; then echo "usage: build <platform>"; exit 1; fi
+
+readonly platform="$1"
+
 rm -rf build
 mkdir build
 cd build
-cmake ..
-make v=1
+cmake .. -DPLATFORM="$platform"
+make
